Move 'addEducation', 'removeEducation', 'handleEducation' to separate files.
