+++ src/components/form/components/education/units/addEducation.js
@@ -1,0 +1,9 @@
+export const addEducation = (resumeData, setResumeData) => {
+  setResumeData({
+    ...resumeData,
+    education: [
+      ...resumeData.education,
+      {school: "", degree: "", startYear: "", endYear: ""},
+    ],
+  });
+};
