+++ src/components/form/components/education/units/removeEducation.js
@@ -1,0 +1,6 @@
+export const removeEducation = (resumeData, setResumeData, index) => {
+  const newEducation = [...resumeData.education];
+  newEducation[index] = newEducation[newEducation.length - 1];
+  newEducation.pop();
+  setResumeData({...resumeData, education: newEducation});
+};
